refactor: resolve logging helper cycle
